feat: 원끼리 충돌시 서로 빗겨가도록 구현

+++ game/game.go
@@ -65,16 +65,16 @@ func (game *Game) SetPlayingStatus(length int) *Game {
 
 	// player relocation
 	game.state.Players.Range(func(key string, player *object.Player) bool {
-		x := float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
-		y := float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		x := float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		y := float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
 		game.calculator.ReLocation(player, x, y)
 		return true
 	})
 
 	// healpack setting
 	for i := 0; i < length*length; i++ {
-		x := float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
-		y := float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		x := float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		y := float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
 		newHealPack := object.NewHealPack(x, y)
 		game.calculator.SetLocation(newHealPack, x, y)
 		game.state.HealPacks.Store(newHealPack.Id, newHealPack)
@@ -82,13 +82,13 @@ func (game *Game) SetPlayingStatus(length int) *Game {
 
 	// magic item setting
 	for i := 0; i < length*length; i++ {
-		x := float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
-		y := float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		x := float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		y := float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
 		newStoneItem := object.NewMagicItem(object.STONE_MAGIC, x, y)
 		game.calculator.SetLocation(newStoneItem, x, y)
 		game.state.MagicItems.Store(newStoneItem.ItemId, newStoneItem)
-		x = float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
-		y = float32(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		x = float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
+		y = float64(rand.Intn(int(game.state.MaxCoord-1))) + 0.5
 		newFireItem := object.NewMagicItem(object.FIRE_MAGIC, x, y)
 		game.calculator.SetLocation(newFireItem, x, y)
 		game.state.MagicItems.Store(newFireItem.ItemId, newFireItem)
